Интегрировал учет совершенных сделок | LuckyCurrencyTest

+++ LuckyCurrencyTest/Services/Models/LastTrade/LastTradeData.cs
@@ -10,9 +10,9 @@ namespace LuckyCurrencyTest.Services.Models.LastTrade
         [JsonProperty("tick_direction")]
         public string Tick_direction { get; set; }
         [JsonProperty("price")]
-        public int Price { get; set; }
+        public double Price { get; set; }
         [JsonProperty("size")]
-        public int Size { get; set; }
+        public double Size { get; set; }
         [JsonProperty("timestamp")]
         public DateTime Timestamp { get; set; }
         [JsonProperty("trade_time_ms")]
